+++ rx-location-manager-nmea/src/test/kotlin/net/samystudio/rxlocationmanager/nmea/GLLTest.kt
@@ -3,11 +3,7 @@ package net.samystudio.rxlocationmanager.nmea
 import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-import org.robolectric.annotation.Config
 
-@RunWith(RobolectricTestRunner::class)
-@Config(manifest = Config.NONE)
 class GLLTest {
     @Test
     fun validate() {