+++ rx-location-manager-nmea/src/test/kotlin/net/samystudio/rxlocationmanager/nmea/LocationTest.kt
@@ -2,12 +2,7 @@ package net.samystudio.rxlocationmanager.nmea
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
-import org.junit.runner.RunWith
-import org.robolectric.RobolectricTestRunner
-import org.robolectric.annotation.Config
 
-@RunWith(RobolectricTestRunner::class)
-@Config(manifest = Config.NONE)
 class LocationTest {
     @Test
     fun convert() {